+++ GoDiveMVP/Data/ActivityUploadRoute.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum ActivityUploadRoute: Hashable, Sendable {
+    case fitImportOptions
+    case uddfImportOptions
+    case macDiveImportGuide
+}